feat: rework entire selection page
